feat(logger): expose add_quebec_context structlog processor
